+++ cgroups/memory.go
@@ -40,7 +40,10 @@ func (m *MemorySubSystem) Apply(cgroupPath string, pid int) error {
 func (m *MemorySubSystem) Remove(cgroupPath string) error {
 	// 删除cgroup相关的文件
 	if subsysCgroupPath, err := GetCgroupPath(m.Name(), cgroupPath, false); err == nil {
-		return os.Remove(subsysCgroupPath)
+		if _, err := os.Stat(subsysCgroupPath); err == nil {
+			return os.Remove(subsysCgroupPath)
+		}
+		return err
 	} else {
 		return err
 	}
